take the amount in string and only convert to int just before calling stripe (#1724)
take the amount in string and only convert to int just before calling stripe

Stripe expects an integer. So $1.23 should be sent as 123. Earlier code was treating amount as int and reducing $1.23 to $1

+++ ypl/backend/payment/stripe/stripe_payout.py
@@ -103,7 +103,7 @@ class StripePayout:
     currency: str
     recipient_account_id: str
     destination_id: str
-    amount: int
+    amount: Decimal
 
 
 class StripeUseCaseType(StrEnum):
